ddasd

+++ SummarizationOptimizer.py
@@ -745,7 +745,7 @@ class Tuna_Swamp_Optimizer:
         epoch_record = []
         for data in range(len(self.df)):
           print("===========================")
-          print(self.fitur.loc[[data]].values.flatten())
+          print(self.fitur.loc[[data]].values.flatten().astype("float"))
           input()
           epoch_record.append(sum(self.fitur.loc[[data]].values.flatten()*w))
